// src/interfaces/stakeholder.interface.ts
export interface Stakeholder {
    id: number;
    stakeholder_code: string;
    stakeholder_name: string;
    stakeholder_address: string;
    stakeholder_ofc_phone: string;
    government_agencies: string;
    non_government_agencies: string;
    other_agency: string;
    contact_person: string;
    contact_mobile: string;
    contact_home_phone: string;
    comment: string;
}